:truck: refactor: Move shortest path solver to GameSolver.Solver.ShortestPath namespace
