--- GameSolver/Solver/ISolver.cs
@@ -1,8 +1,0 @@
-﻿using GameSolver.Core.Action;
-
-namespace GameSolver.Solver;
-
-public interface ISolver
-{
-    IReadOnlyList<IGameAction> Solve();
-}
