+++ GameSolver/Solver/ShortestPath/IShortestPathSolver.cs
@@ -1,0 +1,8 @@
+﻿using GameSolver.Core.Action;
+
+namespace GameSolver.Solver.ShortestPath;
+
+public interface IShortestPathSolver
+{
+    IReadOnlyList<IGameAction> Solve();
+}
